+++ src/app/api/index.ts
@@ -1,13 +1,13 @@
 import { Observable } from 'rxjs/Observable';
 import { normalize } from 'normalizr';
-import uuid from 'uuid/v4';
+import _ from 'lodash';
 
 import { todoListSchema, Todo, NormalizedTodoList } from '../models';
 
 const TODOS: Todo[] = [
-  { id: uuid(), text: 'foo', completed: false },
-  { id: uuid(), text: 'bar', completed: false },
-  { id: uuid(), text: 'baz', completed: false },
+  { id: _.uniqueId(), text: 'foo', completed: false },
+  { id: _.uniqueId(), text: 'bar', completed: false },
+  { id: _.uniqueId(), text: 'baz', completed: false },
 ];
 
 export const getTodos: () => Observable<NormalizedTodoList> = () => {