+++ src/app/reducers/index.ts
@@ -1,4 +1,5 @@
 import { combineReducers } from 'redux';
+import { RouterState } from 'connected-react-router';
 
 import { todosReducer as todos, TodosState } from './todos';
 
@@ -7,7 +8,10 @@ const rootReducer = combineReducers<State>({
 });
 
 export interface State {
-  todos?: TodosState;
+  readonly todos?: TodosState;
+  readonly router?: RouterState;
 }
+
+export const INITIAL_STATE: State = {};
 
 export default rootReducer;