+++ webpack/webpack.dev.js
@@ -46,7 +46,12 @@ module.exports = {
         test: /\.tsx?$/,
         exclude: /node_modules/,
         use: [
-          'babel-loader',
+          {
+            loader: 'babel-loader',
+            options: {
+              cacheDirectory: true,
+            },
+          },
           {
             loader: 'awesome-typescript-loader',
             options: {